refactor: split vlm-inference into generic VLM example and PaddleOCR pipeline
vlm-inference.ts is now a clean, general-purpose VLM inference example
that supports Qianfan-OCR (and any future VLM) with CLI flags:
  --image, --prompt, --stream, --thinking, --max-tokens, --temperature

The PaddleOCR-specific layout+OCR pipeline is moved to
paddle-ocr-pipeline.ts (unchanged functionality).
